+++ js/ui/filter_builder.d.ts
@@ -1,5 +1,4 @@
-import { CommonDataSource } from '../data/data_source_aliases';
-import DataSource from '../data/data_source';
+import DataSource, { CommonDataSource } from '../data/data_source';
 import {
     UserDefinedElement,
     DxElement,
