+++ js/viz/funnel.d.ts
@@ -1,4 +1,4 @@
-import { CommonDataSource } from '../data/data_source_aliases';
+import DataSource, { CommonDataSource } from '../data/data_source';
 
 import {
     UserDefinedElement,
@@ -13,8 +13,6 @@ import {
 import {
     template,
 } from '../core/templates/template';
-
-import DataSource from '../data/data_source';
 
 import {
     EventInfo,
