+++ ts/dx.all.d.ts
@@ -1648,16 +1648,6 @@ declare module DevExpress.data {
     selector: KeySelector<T>;
   };
   /**
-   * [descr:CommonDataSource]
-   * @deprecated Attention! This type is for internal purposes only. If you used it previously, please describe your scenario in the following GitHub Issue, and we will suggest a public alternative: {@link https://github.com/DevExpress/DevExtreme/issues/17885|Internal Types}.
-   */
-  export type CommonDataSource<TItem, TKey = any> =
-    | string
-    | Array<TItem>
-    | Store<TItem, any, TKey>
-    | DevExpress.data.DataSource.Options<any, TItem, any, any, TKey>
-    | DataSource<TItem, any, TKey>;
-  /**
    * [descr:CustomStore]
    */
   export class CustomStore<
@@ -1911,6 +1901,16 @@ declare module DevExpress.data {
     totalCount(): number;
   }
   module DataSource {
+    /**
+     * [descr:CommonDataSource]
+     * @deprecated Attention! This type is for internal purposes only. If you used it previously, please describe your scenario in the following GitHub Issue, and we will suggest a public alternative: {@link https://github.com/DevExpress/DevExtreme/issues/17885|Internal Types}.
+     */
+    export type CommonDataSource<TItem, TKey = any> =
+      | string
+      | Array<TItem>
+      | Store<TItem, any, TKey>
+      | Options<any, TItem, any, any, TKey>
+      | DataSource<TItem, any, TKey>;
     /**
      * @deprecated Attention! This type is for internal purposes only. If you used it previously, please describe your scenario in the following GitHub Issue, and we will suggest a public alternative: {@link https://github.com/DevExpress/DevExtreme/issues/17885|Internal Types}.
      */
@@ -4315,7 +4315,7 @@ declare module DevExpress.ui {
     /**
      * [descr:CollectionWidgetOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<TItem, TKey>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<TItem, TKey>;
     /**
      * [descr:CollectionWidgetOptions.itemHoldTimeout]
      */
@@ -4497,7 +4497,9 @@ declare module DevExpress.ui {
     /**
      * [descr:DataExpressionMixinOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<CollectionWidgetItem | any>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
+      CollectionWidgetItem | any
+    >;
     /**
      * [descr:DataExpressionMixinOptions.displayExpr]
      */
@@ -4661,7 +4663,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxAccordionOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxAccordion.Item | any
     >;
     /**
@@ -4806,7 +4808,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxActionSheetOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxActionSheet.Item | any
     >;
     /**
@@ -4974,7 +4976,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxBoxOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxBox.Item | any
     >;
     /**
@@ -5464,7 +5466,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxContextMenuOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<DevExpress.ui.dxContextMenu.Item>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<DevExpress.ui.dxContextMenu.Item>;
     /**
      * [descr:dxContextMenuOptions.items]
      */
@@ -9657,7 +9659,7 @@ declare module DevExpress.ui {
       /**
        * [descr:dxDiagramOptions.edges.dataSource]
        */
-      dataSource?: DevExpress.data.CommonDataSource<any>;
+      dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
       /**
        * [descr:dxDiagramOptions.edges.fromExpr]
        */
@@ -9788,7 +9790,7 @@ declare module DevExpress.ui {
       /**
        * [descr:dxDiagramOptions.nodes.dataSource]
        */
-      dataSource?: DevExpress.data.CommonDataSource<any>;
+      dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
       /**
        * [descr:dxDiagramOptions.nodes.heightExpr]
        */
@@ -10764,7 +10766,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxDropDownBoxOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<any>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
     /**
      * [descr:dxDropDownBoxOptions.displayValueFormatter]
      */
@@ -10864,7 +10866,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxDropDownButtonOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       DevExpress.ui.dxDropDownButton.Item | any
     >;
     /**
@@ -12145,7 +12147,7 @@ declare module DevExpress.ui {
      * @deprecated Attention! This type is for internal purposes only. If you used it previously, please describe your scenario in the following GitHub Issue, and we will suggest a public alternative: {@link https://github.com/DevExpress/DevExtreme/issues/17885|Internal Types}.
      */
     export type FilterLookupDataSource<T> = Exclude<
-      DevExpress.data.CommonDataSource<T>,
+      DevExpress.data.DataSource.CommonDataSource<T>,
       string | DevExpress.data.DataSource
     >;
     export type InitializedEvent =
@@ -13018,7 +13020,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxGalleryOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxGallery.Item | any
     >;
     /**
@@ -13567,7 +13569,7 @@ declare module DevExpress.ui {
       /**
        * [descr:dxGanttOptions.dependencies.dataSource]
        */
-      dataSource?: DevExpress.data.CommonDataSource<any>;
+      dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
       /**
        * [descr:dxGanttOptions.dependencies.keyExpr]
        */
@@ -13800,7 +13802,7 @@ declare module DevExpress.ui {
       /**
        * [descr:dxGanttOptions.resourceAssignments.dataSource]
        */
-      dataSource?: DevExpress.data.CommonDataSource<any>;
+      dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
       /**
        * [descr:dxGanttOptions.resourceAssignments.keyExpr]
        */
@@ -13825,7 +13827,7 @@ declare module DevExpress.ui {
       /**
        * [descr:dxGanttOptions.resources.dataSource]
        */
-      dataSource?: DevExpress.data.CommonDataSource<any>;
+      dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
       /**
        * [descr:dxGanttOptions.resources.keyExpr]
        */
@@ -13913,7 +13915,7 @@ declare module DevExpress.ui {
       /**
        * [descr:dxGanttOptions.tasks.dataSource]
        */
-      dataSource?: DevExpress.data.CommonDataSource<any>;
+      dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
       /**
        * [descr:dxGanttOptions.tasks.endExpr]
        */
@@ -14335,7 +14337,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxHtmlEditorMention.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<string>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<string>;
     /**
      * [descr:dxHtmlEditorMention.displayExpr]
      */
@@ -14501,7 +14503,8 @@ declare module DevExpress.ui {
    * @deprecated Use DevExpress.ui.dxHtmlEditor.ContextMenuItem instead
    * @deprecated Attention! This type is for internal purposes only. If you used it previously, please describe your scenario in the following GitHub Issue, and we will suggest a public alternative: {@link https://github.com/DevExpress/DevExtreme/issues/17885|Internal Types}.
    */
-  export interface dxHtmlEditorTableContextMenuItem extends MenuBasePlainItem {
+  export interface dxHtmlEditorTableContextMenuItem
+    extends DevExpress.ui.dxMenu.MenuBasePlainItem {
     /**
      * [descr:dxHtmlEditorTableContextMenuItem.name]
      */
@@ -14777,7 +14780,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxHtmlEditorVariables.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<string>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<string>;
     /**
      * [descr:dxHtmlEditorVariables.escapeChar]
      */
@@ -15086,7 +15089,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxListOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxList.Item | any,
       TKey
     >;
@@ -15916,7 +15919,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxMenuBaseOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<dxMenuBaseItem>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<dxMenuBaseItem>;
     /**
      * [descr:dxMenuBaseOptions.items]
      */
@@ -15979,7 +15982,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxMenuOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<DevExpress.ui.dxMenu.Item>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<DevExpress.ui.dxMenu.Item>;
     /**
      * [descr:dxMenuOptions.hideSubmenuOnMouseLeave]
      */
@@ -16091,7 +16094,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxMultiViewOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxMultiView.Item | any
     >;
     /**
@@ -17830,7 +17833,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxResponsiveBoxOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxResponsiveBox.Item | any
     >;
     /**
@@ -18264,7 +18267,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxSchedulerOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<DevExpress.ui.dxScheduler.Appointment>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<DevExpress.ui.dxScheduler.Appointment>;
     /**
      * [descr:dxSchedulerOptions.dateCellTemplate]
      */
@@ -18492,7 +18495,7 @@ declare module DevExpress.ui {
       /**
        * [descr:dxSchedulerOptions.resources.dataSource]
        */
-      dataSource?: DevExpress.data.CommonDataSource<any>;
+      dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
       /**
        * [descr:dxSchedulerOptions.resources.displayExpr]
        */
@@ -19103,7 +19106,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxSlideOutOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxSlideOut.Item | any
     >;
     /**
@@ -19681,7 +19684,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxTabPanelOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxTabPanel.Item | any
     >;
     /**
@@ -19789,7 +19792,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxTabsOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxTabs.Item | any
     >;
     /**
@@ -20291,7 +20294,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxTileViewOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxTileView.Item | any
     >;
     /**
@@ -20490,7 +20493,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxToolbarOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<
       string | DevExpress.ui.dxToolbar.Item | any
     >;
     /**
@@ -22041,7 +22044,7 @@ declare module DevExpress.ui {
     /**
      * [descr:dxTreeViewOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<DevExpress.ui.dxTreeView.Item>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<DevExpress.ui.dxTreeView.Item>;
     /**
      * [descr:dxTreeViewOptions.dataStructure]
      */
@@ -22868,7 +22871,7 @@ declare module DevExpress.ui {
     /**
      * [descr:GridBaseOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<TRowData, TKey>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<TRowData, TKey>;
     /**
      * [descr:GridBaseOptions.dateSerializationFormat]
      */
@@ -23851,7 +23854,7 @@ declare module DevExpress.viz {
     /**
      * [descr:BaseChartOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<any>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
     /**
      * [descr:BaseChartOptions.legend]
      */
@@ -29813,7 +29816,7 @@ declare module DevExpress.viz {
     /**
      * [descr:dxFunnelOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<any>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
     /**
      * [descr:dxFunnelOptions.hoverEnabled]
      */
@@ -32600,7 +32603,7 @@ declare module DevExpress.viz {
     /**
      * [descr:dxRangeSelectorOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<any>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
     /**
      * [descr:dxRangeSelectorOptions.dataSourceField]
      */
@@ -33106,7 +33109,7 @@ declare module DevExpress.viz {
     /**
      * [descr:dxSankeyOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<any>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
     /**
      * [descr:dxSankeyOptions.hoverEnabled]
      */
@@ -33499,7 +33502,7 @@ declare module DevExpress.viz {
     /**
      * [descr:dxSparklineOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<any>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
     /**
      * [descr:dxSparklineOptions.firstLastColor]
      */
@@ -33781,7 +33784,7 @@ declare module DevExpress.viz {
     /**
      * [descr:dxTreeMapOptions.dataSource]
      */
-    dataSource?: DevExpress.data.CommonDataSource<any>;
+    dataSource?: DevExpress.data.DataSource.CommonDataSource<any>;
     /**
      * [descr:dxTreeMapOptions.group]
      */
@@ -34385,7 +34388,9 @@ declare module DevExpress.viz {
           /**
            * [descr:dxVectorMapOptions.layers.dataSource]
            */
-          dataSource?: object | DevExpress.data.CommonDataSource<any>;
+          dataSource?:
+            | object
+            | DevExpress.data.DataSource.CommonDataSource<any>;
           /**
            * [descr:dxVectorMapOptions.layers.elementType]
            */
@@ -34492,7 +34497,9 @@ declare module DevExpress.viz {
           colorGroups?: Array<number>;
           customize?: (elements: Array<MapLayerElement>) => any;
           dataField?: string;
-          dataSource?: object | DevExpress.data.CommonDataSource<any>;
+          dataSource?:
+            | object
+            | DevExpress.data.DataSource.CommonDataSource<any>;
           elementType?: 'bubble' | 'dot' | 'image' | 'pie';
           hoverEnabled?: boolean;
           hoveredBorderColor?: string;
